services/ai_service.py
from sklearn.naive_bayes import MultinomialNB
from sklearn.feature_extraction.text import CountVectorizer
import numpy as np

# Dummy data for learning style classification
learning_styles = ['visual', 'auditory', 'kinesthetic', 'reading/writing']
sample_responses = [
    "I prefer diagrams and charts",
    "I learn best through lectures and discussions",
    "Hands-on activities help me understand better",
    "I enjoy reading textbooks and writing notes"
]
sample_labels = [0, 1, 2, 3]  # Corresponding to learning_styles

# Train a simple Naive Bayes classifier
vectorizer = CountVectorizer()
X = vectorizer.fit_transform(sample_responses)
clf = MultinomialNB()
clf.fit(X, sample_labels)

def assess_learning_style(questionnaire_data):
    # Process questionnaire data
    responses = [q['answer'] for q in questionnaire_data]
    X_new = vectorizer.transform(responses)
    
    # Predict learning style
    predictions = clf.predict(X_new)
    
    # Get the most common prediction
    learning_style_index = np.bincount(predictions).argmax()
    return learning_styles[learning_style_index]

def adapt_content_difficulty(user_performance):
    # Implement adaptive difficulty logic
    if user_performance > 0.8:
        return 'hard'
    elif user_performance > 0.5:
        return 'medium'
    else:
        return 'easy'
